move Apt & Tar funcs off command
