+++ pkg/docker/api_client.go
@@ -219,14 +219,7 @@ func (c *apiClient) Push(ctx context.Context, imageRef string) error {
 // TODO[md]: this doesn't need to be on the interface, move to auth handler
 func (c *apiClient) LoadUserInformation(ctx context.Context, registryHost string) (*command.UserInfo, error) {
 	console.Debugf("=== APIClient.LoadUserInformation %s", registryHost)
-}
 
-func (c *apiClient) CreateTarFile(ctx context.Context, ref string, tmpDir string, tarFile string, folder string) (string, error) {
-	panic("not implemented")
-}
-
-func (c *apiClient) CreateAptTarFile(ctx context.Context, tmpDir string, aptTarFile string, packages ...string) (string, error) {
-	panic("not implemented")
 	return loadUserInformation(ctx, registryHost)
 }
